Allow trees containing many children
Previously, if there were many (~100k+, depending on the browser) children of a tree node, this code would fail with an error about exceeding the maximum stack size. This is because we were attempting to pass too many arguments to Array.push. Instead, naively iterate through the array of descendants and add them one-by-one.

Signed-off-by: Carl Jackson <carl@avtok.com>