Make checkScaleErrors accept less strict type
This allows to remove a ts-expect-error in checkScaleErrors.test.ts which is hiding some more type violations in the same function call below it.

I chose to remove the DefaultizedBarSeriesType type constraint because the checkScaleError never reads properties from there - it is only concerned with the stackedData array.

Alternatively I could write a test helper that returns the required properties from DefaultizedBarSeriesType and use that in the test - let me know if you prefer that instead.